refactor ThreadResult
