+++ app/src/main/java/com/infomaniak/mail/data/models/thread/ThreadResult.kt
@@ -1,0 +1,31 @@
+/*
+ * Infomaniak kMail - Android
+ * Copyright (C) 2023 Infomaniak Network SA
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.infomaniak.mail.data.models.thread
+
+import com.infomaniak.mail.data.models.Thread
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ThreadResult(
+    val threads: List<Thread>? = null,
+    @SerialName("resource_previous")
+    val resourcePrevious: String?,
+    @SerialName("resource_next")
+    val resourceNext: String?,
+)
